#240 Send redirect
