+++ src/main/java/com/cloudogu/smeagol/authc/infrastructure/CasInfrastructureRegistration.java
@@ -22,6 +22,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import java.io.IOException;
+import java.net.URL;
 import java.util.Arrays;
 import java.util.stream.Collectors;
 
@@ -78,37 +79,7 @@ public class CasInfrastructureRegistration {
      */
     @Bean
     public FilterRegistrationBean proxyReceivingTicketValidationFilter() {
-        return casFilterRegistration(new Cas30ProxyReceivingTicketValidationFilterWithoutException(), 1);
-    }
-
-    public static class Cas30ProxyReceivingTicketValidationFilterWithoutException extends Cas30ProxyReceivingTicketValidationFilter {
-        public Cas30ProxyReceivingTicketValidationFilterWithoutException() {
-            super();
-        }
-
-        @Override
-        protected void onFailedValidation(HttpServletRequest request, HttpServletResponse response) {
-            try {
-                final String requestUrl = request.getRequestURL().toString();
-                final String queryString = request.getQueryString();
-
-                String newQuery = "";
-                if (queryString != null) {
-                    newQuery = Arrays.stream(queryString.split("&"))
-                        .filter(param -> !param.startsWith("ticket="))
-                        .collect(Collectors.joining("&"));
-                }
-
-                final String newUrl = requestUrl + (newQuery.isEmpty() ? "" : "?" + newQuery);
-                System.out.println("=================>>");
-                System.out.println("!=================>>");
-                System.out.println("=================>>");
-                System.out.println(newUrl);
-                response.sendRedirect(newUrl);
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-        }
+        return casFilterRegistration(new CESCas30ProxyReceivingTicketValidationFilter(), 1);
     }
 
     /**
