+++ src/main/java/com/cloudogu/smeagol/authc/infrastructure/TicketEmptyException.java
@@ -1,0 +1,4 @@
+package com.cloudogu.smeagol.authc.infrastructure;
+
+public class TicketEmptyException extends RuntimeException{
+}
